Image handling, helper functions, layout improvements
Used a bit of a brute-force approach here for image URL formatting. Implemented as both an Image.svelte component (needs additional props to be fully functional) and as a utility function (currently co-located with the sanity-specific components).

the `groq` package adds a no-op template string function, that's used as a syntax highlighting entrypoint by sanity's VSCode extension.